+++ src/clj/py99/routes/home.clj
@@ -249,18 +249,31 @@
        (expand-includes (get-answer (Integer/parseInt num) login) login)
        line))))
 
+;; 2023-10-19
+(defn- has-docstring-test
+  "if s contains docstring returns nil or throw.
+   FIXME: should check `def` proceeds the comment line."
+  [lines]
+  (if (some
+       (fn [s] (or (re-find #"^\s+\"\"\"" s)
+                   (re-find #"^\s+\s\'\'\'" s)))
+       (str/split-lines lines))
+    nil
+    (throw (Exception. "no docstring"))))
+
 (defn- validate
   "Return nil if all validations success, or raize exeption."
   [num answer login]
   (try
     (not-empty-test (strip answer))
+    (has-docstring-test answer)
     (pytest-test num (expand-includes answer login))
     nil
     (catch Exception e (throw (Exception. (.getMessage e))))))
 
 (defn create-answer!
   [{{:keys [num answer]} :params :as request}]
-  ;; (log/debug "create-answer!" (login request) num)
+  (log/debug "create-answer!" (login request) num)
   (try
     (when-not (env :exam-mode)
       (validate (Integer/parseInt num) answer (login request)))
@@ -269,7 +282,10 @@
       :num (Integer/parseInt num)
       :answer answer
       :md5 (-> answer strip digest/md5)})
-    (redirect (str "/answer/" num))
+    ;; 2023-10-15
+    (if (env :dev)
+      (redirect (str "/answer/" num))
+      (redirect "https://qa.melt.kyutech.ac.jp/qs"))
     (catch Exception e
       (layout/render request "error.html"
                      {:status 406
@@ -306,7 +322,7 @@
 (defn create-comment! [request]
   (let [params (:params request)
         num (Integer/parseInt (:p_num params))]
-    (log/debug "create-comment! num:" num)
+    (log/debug "create-comment!" (login request) num)
     (if (db/frozen? {:num num})
       (layout/render request "error.html"
                      {:status 403